+++ native.js
@@ -1,0 +1,4 @@
+var PhoneInput = require('./commonjs/PhoneInputNativeDefaultMetadata').default
+
+exports = module.exports = PhoneInput
+exports['default']       = PhoneInput