+++ react-responsive-ui.js
@@ -1,0 +1,4 @@
+var PhoneInput = require('./commonjs/PhoneInputReactResponsiveUIDefaultMetadata').default
+
+exports = module.exports = PhoneInput
+exports['default']       = PhoneInput