+++ source/PhoneInputNative.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import classNames from 'classnames'
 
 import PhoneInput from './PhoneInput'
 import CountrySelect from './CountrySelectNative'
@@ -13,7 +14,21 @@ export default class PhoneInputNative extends Component
 			<PhoneInput
 				{ ...this.props }
 				ref={ this.storeInputRef }
-				countrySelectComponent={ CountrySelect }/>
+				countrySelectComponent={ CountrySelect }
+				getInputClassName={ this.getInputClassName }/>
+		)
+	}
+
+	getInputClassName = ({ disabled, invalid }) =>
+	{
+		return classNames
+		(
+			// Will be removed in version 2.x
+			'react-phone-number-input__phone--native',
+			{
+				'react-phone-number-input__phone--disabled' : disabled,
+				'react-phone-number-input__phone--invalid'  : invalid
+			}
 		)
 	}