+++ source/PhoneInputReactResponsiveUI.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import classNames from 'classnames'
 
 import PhoneInput from './PhoneInput'
 import CountrySelect from './CountrySelectReactResponsiveUI'
@@ -13,7 +14,22 @@ export default class PhoneInputReactResponsiveUI extends Component
 			<PhoneInput
 				{ ...this.props }
 				ref={ this.storeInputRef }
-				countrySelectComponent={ CountrySelect }/>
+				countrySelectComponent={ CountrySelect }
+				getInputClassName={ this.getInputClassName }/>
+		)
+	}
+
+	getInputClassName = ({ disabled, invalid }) =>
+	{
+		return classNames
+		(
+			'rrui__input',
+			'rrui__input-element',
+			'rrui__input-field',
+			{
+				'rrui__input-field--invalid'  : invalid,
+				'rrui__input-field--disabled' : disabled
+			}
 		)
 	}